server/requirements.txt
aniso8601==7.0.0
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.4
dataclasses==0.6
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.40.0
django-cors-headers==3.1.0
django-filter==2.2.0
django-graphiql==0.4.4
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djongo==1.2.35
graphene==2.1.8
graphene-django==2.5.0
graphql-core==2.2.1
graphql-relay==2.0.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
oauthlib==3.1.0
pep8==1.7.1
Pillow==6.1.0
pkg-resources==0.0.0
promise==2.2.1
PyJWT==1.7.1
pylint==2.3.1
pymongo==3.9.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.2.4
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
